# swagger.yaml
openapi: 3.0.0
info:
  title: SquidFlow Platform
  version: '1.0.0'
  description: API for managing applications, clusters, templates and secrets
...
servers:
  - url: http://localhost:38080/api/v1
    description: Local development server
components:
  schemas:
...
    SecretStoreCreate:
      type: object
      required:
        - secret_store_yaml
      properties:
        secret_store_yaml:
          type: string
          description: YAML representation of the SecretStore configuration
    SecretStoreCreateResponse:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        success:
          type: boolean
        message:
          type: string
    SecretStoreInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        type:
          type: string
        status:
          type: string
        path:
          type: string
        lastSynced:
          type: string
        createdAt:
          type: string
        lastUpdated:
          type: string
        health:
          type: object
          properties:
            status:
              type: string
            message:
              type: string
    SecretStoreUpdate:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        server:
          type: string
        auth:
          type: object
          additionalProperties: true
...
paths:
...
  /security/externalsecrets/secretstore:
    post:
      tags: [Security]
      summary: Create SecretStore
      operationId: SecretStoreCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretStoreCreate'
      responses:
        '201':
          description: SecretStore created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretStoreCreateResponse'

    get:
      tags: [Security]
      summary: List SecretStores
      operationId: SecretStoreList
      responses:
        '200':
          description: List of SecretStores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretStoreInfo'
  /security/externalsecrets/secretstore/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [Security]
      summary: Get SecretStore details
      operationId: SecretStoreDescribe
      responses:
        '200':
          description: SecretStore details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretStoreInfo'
        '404':
          description: SecretStore not found

    delete:
      tags: [Security]
      summary: Delete SecretStore
      operationId: SecretStoreDelete
      responses:
        '200':
          description: SecretStore deleted successfully
        '404':
          description: SecretStore not found

    patch:
      tags: [Security]
      summary: Update SecretStore
      operationId: SecretStoreUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretStoreUpdate'
      responses:
        '200':
          description: SecretStore updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretStoreInfo'
        '404':
          description: SecretStore not found
...